pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soplex"
version = "0.1.0"
description = "Compile plain-English SOPs into executable, cost-optimized agent graphs"
authors = [
    {name = "soplex", email = "info@soplex.dev"}
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["llm", "agents", "automation", "sop", "workflow", "ai"]

dependencies = [
    "openai>=1.40.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.34.0"]
litellm = ["litellm>=1.40.0"]
all = ["anthropic>=0.34.0", "litellm>=1.40.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.5.0",
    "mypy>=1.8.0",
]

[project.scripts]
soplex = "soplex.cli:app"

[project.urls]
Documentation = "https://github.com/soplex-ai/soplex"
Repository = "https://github.com/soplex-ai/soplex"
Homepage = "https://soplex.dev"

[tool.hatch.build.targets.sdist]
include = [
    "/src/soplex",
    "/examples",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["src/soplex"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: marks tests as end-to-end (may make real API calls)",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
target-version = "py39"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "ARG", "SIM"]
ignore = ["E501", "W503", "E203"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true